refactor `AeCiphertext` into a separate submodule
