+++ zk-token-sdk/src/zk_token_elgamal/pod/auth_encryption.rs
@@ -1,0 +1,33 @@
+use {
+    crate::zk_token_elgamal::pod::{Pod, Zeroable},
+    base64::{prelude::BASE64_STANDARD, Engine},
+    std::fmt,
+};
+
+/// Serialization for AeCiphertext
+#[derive(Clone, Copy, PartialEq, Eq)]
+#[repr(transparent)]
+pub struct AeCiphertext(pub [u8; 36]);
+
+// `AeCiphertext` is a Pod and Zeroable.
+// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
+unsafe impl Zeroable for AeCiphertext {}
+unsafe impl Pod for AeCiphertext {}
+
+impl fmt::Debug for AeCiphertext {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{:?}", self.0)
+    }
+}
+
+impl fmt::Display for AeCiphertext {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", BASE64_STANDARD.encode(self.0))
+    }
+}
+
+impl Default for AeCiphertext {
+    fn default() -> Self {
+        Self::zeroed()
+    }
+}
