+++ zk-token-sdk/src/zk_token_elgamal/pod/mod.rs
@@ -1,3 +1,4 @@
+mod auth_encryption;
 mod elgamal;
 mod pedersen;
 mod range_proof;
@@ -5,12 +6,11 @@ mod sigma_proofs;
 
 use {
     crate::zk_token_proof_instruction::ProofType,
-    base64::{prelude::BASE64_STANDARD, Engine},
     num_traits::{FromPrimitive, ToPrimitive},
     solana_program::instruction::InstructionError,
-    std::fmt,
 };
 pub use {
+    auth_encryption::AeCiphertext,
     bytemuck::{Pod, Zeroable},
     elgamal::{ElGamalCiphertext, ElGamalPubkey},
     pedersen::{DecryptHandle, PedersenCommitment},
@@ -69,34 +69,6 @@ impl TryFrom<PodProofType> for ProofType {
 #[derive(Clone, Copy, Pod, Zeroable, PartialEq, Eq)]
 #[repr(transparent)]
 pub struct CompressedRistretto(pub [u8; 32]);
-
-/// Serialization for AeCiphertext
-#[derive(Clone, Copy, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct AeCiphertext(pub [u8; 36]);
-
-// `AeCiphertext` is a Pod and Zeroable.
-// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
-unsafe impl Zeroable for AeCiphertext {}
-unsafe impl Pod for AeCiphertext {}
-
-impl fmt::Debug for AeCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
-impl fmt::Display for AeCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl Default for AeCiphertext {
-    fn default() -> Self {
-        Self::zeroed()
-    }
-}
 
 // TODO: refactor this code into the instruction module
 #[derive(Clone, Copy, Pod, Zeroable)]
